moved role required to auth
